+++ .gitignore
@@ -242,3 +242,5 @@ cython_debug/
 marimo/_static/
 marimo/_lsp/
 __marimo__/
+
+tests/results
