+++ tests/utils/__init__.py
@@ -1,2 +1,5 @@
-from utils.answer_parser import *
-from utils.similarity import *
+from utils.answer_parser import extract_answer_from_output
+from utils.similarity import SimilarityScorer
+from utils.generate_report import generate_summary_report
+
+__all__ = ["SimilarityScorer", "extract_answer_from_output", "generate_summary_report"]
